// src/Button/Button.enums.ts
export enum ButtonWeightEnum {
    SMALL = 'small',
    BIG = 'big',
    DEFAULT = 'big',
}

export enum ButtonVariantsEnum {
    OUTLINED = "outlined",
    DEFAULT = "default",
    TEXT = "text",
}
export enum ButtonColorsEnum {
    PRIMARY = "primary",
    DEFAULT = "default",
    SECONDARY = "secondary",
}

const ButtonEnums = {
    ButtonWeightEnum,
    ButtonVariantsEnum,
    ButtonColorsEnum,
}

export default ButtonEnums;